+++ src/traceml/reporting/summaries/system.py
@@ -18,12 +18,12 @@ import sqlite3
 from dataclasses import dataclass
 from typing import Any, Dict, Optional
 
-from traceml.aggregator.summaries.summary_formatting import (
+from traceml.reporting.summaries.summary_formatting import (
     bytes_to_gb,
     duration_from_bounds,
     format_optional,
 )
-from traceml.aggregator.summaries.summary_io import (
+from traceml.reporting.summaries.summary_io import (
     load_json_or_empty,
     write_json,
 )
