+++ tests/test_init_and_wrappers.py
@@ -36,9 +36,9 @@ def test_init_auto_enables_all_supported_patches(monkeypatch):
 
     calls = []
 
-    import traceml.utils.patches.backward_auto_timer_patch as backward_patch
-    import traceml.utils.patches.dataloader_patch as dataloader_patch
-    import traceml.utils.patches.forward_auto_timer_patch as forward_patch
+    import traceml.patches.backward_auto_timer_patch as backward_patch
+    import traceml.patches.dataloader_patch as dataloader_patch
+    import traceml.patches.forward_auto_timer_patch as forward_patch
 
     monkeypatch.setattr(
         dataloader_patch,
@@ -81,9 +81,9 @@ def test_init_selective_only_installs_requested_patches(monkeypatch):
 
     calls = []
 
-    import traceml.utils.patches.backward_auto_timer_patch as backward_patch
-    import traceml.utils.patches.dataloader_patch as dataloader_patch
-    import traceml.utils.patches.forward_auto_timer_patch as forward_patch
+    import traceml.patches.backward_auto_timer_patch as backward_patch
+    import traceml.patches.dataloader_patch as dataloader_patch
+    import traceml.patches.forward_auto_timer_patch as forward_patch
 
     monkeypatch.setattr(
         dataloader_patch,
@@ -149,7 +149,7 @@ def test_init_custom_alias_maps_to_selective(monkeypatch):
 
     calls = []
 
-    import traceml.utils.patches.forward_auto_timer_patch as forward_patch
+    import traceml.patches.forward_auto_timer_patch as forward_patch
 
     monkeypatch.setattr(
         forward_patch,
